pyproject.toml
[project]
name = "ai-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Sangwon Lee", email = "lsw91@lgchem.com" }
]
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.51.0",
    "mcp>=1.8.1",
    "notebook>=7.4.2",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "streamlit>=1.45.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
